Validate properties in injection token and add rowheight in O_GLOBAL_CONFIG
