+++ projects/ontimize-web-ngx/src/lib/types/o-global-config.type.ts
@@ -1,7 +1,9 @@
 import { InjectionToken } from "@angular/core";
+import { ORowHeight } from "../util/codes";
 
 export type OGlobalConfig = {
   storeState: boolean;
+  rowHeight: ORowHeight;
 };
 
 export const O_GLOBAL_CONFIG = new InjectionToken<OGlobalConfig>('o-global-config');
