+++ projects/ontimize-web-ngx/src/lib/util/codes.ts
@@ -25,10 +25,12 @@ export class Codes {
   public static DETAIL_MODE_NONE = 'none';
   public static DETAIL_MODE_CLICK = 'click';
   public static DETAIL_MODE_DBLCLICK_VALUES = ['dblclick', 'doubleclick'];
+  static AVAILABLE_DETAIL_MODES = [Codes.DETAIL_MODE_NONE, Codes.DETAIL_MODE_CLICK, ...Codes.DETAIL_MODE_DBLCLICK_VALUES];
 
   public static EDITION_MODE_NONE = 'none';
   public static EDITION_MODE_CLICK = 'click';
   public static EDITION_MODE_DBLCLICK_VALUES = ['dblclick', 'doubleclick'];
+  public static AVAILABLE_MODES = [Codes.EDITION_MODE_NONE, Codes.EDITION_MODE_CLICK, ...Codes.EDITION_MODE_DBLCLICK_VALUES];
 
   public static SELECTION_MODE_NONE = 'none';
   public static SELECTION_MODE_SINGLE = 'single';
@@ -127,6 +129,7 @@ export class Codes {
   public static DEFAULT_CHANGE_EVENT: CHANGE_EVENTS = 'onValueChange';
   public static TYPES_DATE_GROUPS = ["YEAR", "MONTH", "YEAR_MONTH", "YEAR_MONTH_DAY"];
 
+
   static isDoubleClickMode(value: string): boolean {
     return Codes.DETAIL_MODE_DBLCLICK_VALUES.indexOf(value) !== -1;
   }
@@ -161,4 +164,11 @@ export class Codes {
     }
     return !((e.keyCode < 48 || e.keyCode > 57) && (e.keyCode < 96 || e.keyCode > 105));
   }
+  static isValidEditionMode(editionMode: OTableEditionMode) {
+    return Codes.AVAILABLE_MODES.indexOf(editionMode) !== -1;
+  }
+  static isValidDetailMode(detailMode: OTableDetailMode): any {
+    return Codes.AVAILABLE_DETAIL_MODES.indexOf(detailMode) !== -1;
+  }
+
 }
